Forgot to add the headers
the file transfear plugin is changed to hopefully show more exact data



git-svn-id: https://wokjab.svn.sourceforge.net/svnroot/wokjab/wokjab/trunk/wokjab@74 a4058eb7-c820-0410-9b0b-94ec00e1a786

+++ plugins/jep0096-Filetransfear/src/filetransfear.cc
@@ -394,7 +394,7 @@ jep96::PrettySize(unsigned long long size)
 	}
 	
 	std::stringstream msg;
-	msg << double(size)/divider << ending;
+	msg << double(size)/double(divider) << ending;
 	
 	return msg.str();
 }
